mdjourney-refactored/mdjourney-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.12.0
python-multipart==0.0.6
# Core dependencies for configuration and schema management
pyyaml>=6.0.0
jsonschema>=4.0.0
# File system monitoring
watchdog>=3.0.0
# File metadata extraction (optional, will fallback to basic scanner if not available)
# dirmeta>=0.1.0
# HTTP client for gateway
httpx>=0.24.0
# Session management
starlette>=0.27.0
